cloud-functions/email-notifier/src/requirements.txt
sendgrid
google-cloud-firestore
google-cloud-secret-manager